solve conflicts
