--- web-client/src/shared/redux/driverRegistration/index.ts
@@ -1,56 +1,0 @@
-import { createSlice, PayloadAction } from '@reduxjs/toolkit';
-
-export interface PreScreen {
-    policeRecordsCheckOptions: string,
-    policeRecordsCheckDate: string,
-    policeRecordsCheckType: string,
-    drivingAbstractOptions: string,
-    drivingAbstractDate: string,
-    LicenseAndVehicle: {
-        haveCar: boolean,
-        haveGLicense: boolean,
-        have1MInsurance: boolean,
-        have2MInsurance: boolean,
-    }
-
-    LicenseClasses: string
-}
-
-const initPreScreen: PreScreen = {
-    policeRecordsCheckOptions: '',
-    policeRecordsCheckDate: '',
-    policeRecordsCheckType: '',
-    drivingAbstractOptions: '',
-    drivingAbstractDate: '',
-    LicenseAndVehicle: {
-        haveCar: false,
-        haveGLicense: false,
-        have1MInsurance: false,
-        have2MInsurance: false,
-    },
-    LicenseClasses: 'None'
-}
-
-interface RegistrationState {
-    preScreen: PreScreen
-}
-
-
-
-const initialState = {
-    preScreen: initPreScreen
-} as RegistrationState;
-
-const driversRegistrationSlice = createSlice({
-    name: 'driversRegistration',
-    initialState: initialState,
-    reducers: {
-        setPreScreen: (state, { payload }: PayloadAction<PreScreen>) => {
-            state.preScreen = payload;
-        },
-    }
-});
-
-export const { setPreScreen } = driversRegistrationSlice.actions;
-
-export default driversRegistrationSlice;
